wip fix int to double

+++ sources/setup.h
@@ -6,7 +6,7 @@
 /*   By: seblin <seblin@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/08 07:55:53 by svidot            #+#    #+#             */
-/*   Updated: 2024/01/15 12:46:31 by seblin           ###   ########.fr       */
+/*   Updated: 2024/01/19 11:19:40 by seblin           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,9 +20,9 @@
 
 typedef struct s_point
 {
-	int	init_vect[MTX];
-	int	new_vect[MTX];
-	int	line;
+	double	init_vect[MTX];
+	double	new_vect[MTX];
+	int		line;
 }	t_point;
 
 #endif
